Removed support for lambda compression
The primary reason for removing this was that it conflicted with other
resugaring logic.  For example, suppose that you had this `annah` expression:

    \(a : *) -> {0|}

This would desugar to:

    \(a : *) -> \(t : *) -> forall (Sum : *) -> forall (MkSum : t -> Sum) -> Sum

... but then would resugar to:

    \(a : *) (t : *) -> {0|t}

The multilambda would capture the bound `(t : *)` and then prevent it from
being resugared into a sum type section.

The secondary reason I got rid of this was that it made the code too hard to
read, since it became incredibly difficult to visually parse where each bound
argument began and ended when the arguments had very large types.