Changed dynamic linking to use a list instead of a hashmap
The purpose of this is to use `Eq` for comparison instead of `Hashable`.  The
problem is that `Hashable` does not detect alpha equality.

This means that the lookup is much slower, but in the future we may be able to
retrieve the use of `Hashable` by converting to a nameless representation.